[front] Extract model tool rendering helpers
